Simplify the processor baseclass (remove abstraction leaks from PDF, removed unused code)
