+++ src/mmore/process/processors/eml_processor.py
@@ -4,9 +4,9 @@ import email
 from email import policy
 from PIL import Image
 from typing import Dict, Any
-from src.mmore.process.utils import clean_text, create_sample
+from src.mmore.process.utils import clean_text
 from src.mmore.type import FileDescriptor
-from .processor import Processor, ProcessorConfig
+from .processor import Processor, ProcessorConfig, ProcessorResult
 
 logger = logging.getLogger(__name__)
 
@@ -28,8 +28,7 @@ class EMLProcessor(Processor):
         """
         super().__init__(files, config=config or ProcessorConfig())
     
-    @classmethod
-    def accepts(cls, file: FileDescriptor) -> bool:
+    def accepts(self, file: FileDescriptor) -> bool:
         """
         Args:
             file (FileDescriptor): The file descriptor to check.
@@ -44,9 +43,9 @@ class EMLProcessor(Processor):
         Returns:
             tuple: A tuple (False, False) indicating no GPU requirement for both standard and fast modes.
         """
-        return False, False
+        return False
 
-    def process_implementation(self, file_path: str) -> Dict[str, Any]:
+    def process_one_file(self, file_path: str, fast: bool = False) -> ProcessorResult:
         """
         Process a single EML file. Extracts text content, email headers, and embedded images. 
 
@@ -59,12 +58,14 @@ class EMLProcessor(Processor):
         The method parses the EML file, extracts email headers, text content, and embedded images.
         Embedded images are replaced with a placeholder tag from the processor configuration.
         """
+        super().process_one_file(file_path, fast=fast)
+
         try:
             with open(file_path, 'rb') as f:
                 msg = email.message_from_bytes(f.read(), policy=policy.default)
         except Exception as e:
             logger.error(f"Failed to open EML file {file_path}: {e}")
-            return create_sample([], [], file_path)
+            return self.create_sample([], [], file_path)
 
         all_text = []
         embedded_images = []
@@ -99,4 +100,4 @@ class EMLProcessor(Processor):
                 except Exception as e:
                     logger.error(f"Error extracting image from EML: {e}")
 
-        return create_sample(all_text, embedded_images, file_path)
+        return self.create_sample(all_text, embedded_images, file_path)
