+++ src/mmore/process/processors/pdf_processor.py
@@ -9,24 +9,17 @@ import torch.multiprocessing as mp
 import logging
 import os
 from multiprocessing import Pool, cpu_count
-from PIL import Image,  UnidentifiedImageError
+from PIL import Image, UnidentifiedImageError
 from typing import List, Tuple, Any, Dict
 from . import md_processor
 import re
 from marker.models import load_all_models
 import tempfile
 from marker.settings import settings
-from src.mmore.type import FileDescriptor
-from .processor import Processor, ProcessorConfig
+from src.mmore.type import FileDescriptor, MultimodalSample, MultimodalRawInput
+from .processor import Processor, ProcessorConfig, ProcessorResult
 
-from src.mmore.process.utils import (
-    clean_text,
-    create_sample,
-    create_sample_list,
-    merge_split_with_full_page_indexer,
-    create_sample_list_already_saved_images,
-    clean_image,
-)
+from src.mmore.process.utils import clean_text, clean_image
 
 BATCH_SIZE = 1
 
@@ -61,9 +54,6 @@ class PDFProcessor(Processor):
         process_fast_implementation(file_path: str) -> dict:
             A faster, less resource-intensive method to process PDF files, can struggle with scanned documents.
 
-        split_files_across_gpus() -> List[List[FileDescriptor]]:
-            Splits the input files across multiple GPUs for distributed processing.
-
         get_file_len(file: FileDescriptor) -> int:
             Returns the number of pages in a given PDF file.
 
@@ -76,12 +66,12 @@ class PDFProcessor(Processor):
         - It supports distributed processing in multi-GPU clusters.
         - Extracted content is compatible with downstream tasks, such as multimodal indexing and RAG pipelines.
     """
+
     def __init__(self, files, config=None):
         super().__init__(files, config=config or ProcessorConfig())
         self.ocr_models = {device: None for device in range(torch.cuda.device_count())}
 
-    @classmethod
-    def accepts(cls, file: FileDescriptor) -> bool:
+    def accepts(self, file: FileDescriptor) -> bool:
         return file.file_extension.lower() == ".pdf"
 
     def require_gpu(self) -> Tuple[bool, bool]:
@@ -131,7 +121,12 @@ class PDFProcessor(Processor):
             settings.PAGE_SEPARATOR = "0110001001101100011001010110001101101111011001010111010101110010001000000110000101101110011001000010000001100001011100110110000101101100011011000110100101101110011001010110111000100000011101110110010101110010011001010010000001101000011001010111001001100101"  # Trust us
         return self.ocr_models[device_index]
 
-    def process_fast_implementation(self, file_path: str) -> dict:
+    def process_one_file(self, file_path: str, fast: bool = False) -> ProcessorResult:
+        if fast:
+            return self.fast_and_furious(file_path)
+        return self.slow_and_buggy(file_path)
+
+    def fast_and_furious(self, file_path: str, fast: bool = False) -> ProcessorResult:
         pdf_doc = fitz.open(file_path)
         extracted_text = []
         embedded_images = []
@@ -143,13 +138,14 @@ class PDFProcessor(Processor):
 
             for img_info in page.get_images(full=False):
                 image = self._extract_image_from_pdf(pdf_doc, img_info[0])
-                if clean_image(image): # clean image filters images below size 512x512 and variance below 100, these are defaults and can be changed 
+                if clean_image(
+                        image):  # clean image filters images below size 512x512 and variance below 100, these are defaults and can be changed
                     embedded_images.append(image)
                     extracted_text.append(self.config.attachment_tag)
-                    
-        return create_sample(extracted_text, embedded_images)
 
-    def process_implementation(self, file_path: str, temp_dir: str = "tmp/") -> dict:
+        return self.create_sample(extracted_text, embedded_images, file_path, file_path)
+
+    def slow_and_buggy(self, file_path: str, temp_dir: str = "tmp/") -> ProcessorResult:
         def extract_image_in_page(page, current_image_index) -> Tuple[List[str], int]:
             page_images = []
             num_image_in_page = page.count(self.config.attachment_tag)
@@ -201,7 +197,7 @@ class PDFProcessor(Processor):
         index = self.full_page_indexer[torch.cuda.current_device()]
 
         split_pages = split_pages[1:] if split_pages[0] == "{0}" else split_pages
-        
+
         all_files = []
         all_images = []
         IMAGE_LIST = list(images.values())
@@ -212,7 +208,7 @@ class PDFProcessor(Processor):
 
         previous_file_name = None
         current_image_index = 0
-        
+
         for page, (true_page_number, filename) in zip(split_pages, index):
             if previous_file_name is None:
                 previous_file_name = filename
@@ -247,10 +243,11 @@ class PDFProcessor(Processor):
                 new_results.append(data)
                 break
             last_filename = results[i][-1]['metadata']['file_path']
-            if results[i+1][0]['metadata']['file_path'] == last_filename:
-                need_to_append = results[i+1][0]
-                logger.info(f"File {results[i+1][0]['metadata']['file_path']} was splitted across GPUs and the results are inconsistent. We need to reconstruct the results.")
-                results[i+1] = results[i+1][1:]
+            if results[i + 1][0]['metadata']['file_path'] == last_filename:
+                need_to_append = results[i + 1][0]
+                logger.info(
+                    f"File {results[i + 1][0]['metadata']['file_path']} was splitted across GPUs and the results are inconsistent. We need to reconstruct the results.")
+                results[i + 1] = results[i + 1][1:]
                 temp = results[i]
                 last_sample = results[i][-1]
                 last_sample['text'] += need_to_append['text']
@@ -260,39 +257,119 @@ class PDFProcessor(Processor):
                 new_results.append(results[i])
         return new_results
 
-
     @staticmethod
     def _extract_image_from_pdf(pdf_doc, xref) -> Image.Image:
         try:
             base_image = pdf_doc.extract_image(xref)
             image_bytes = base_image.get("image")
-            
+
             if image_bytes is None:
-                logging.error(f"No image data found for xref {xref}") 
+                logging.error(f"No image data found for xref {xref}")
 
             return Image.open(io.BytesIO(image_bytes)).convert("RGB")
-        
+
         except KeyError as e:
             logging.error(f"KeyError while extracting image: {e}")
             return None
-        
+
         except UnidentifiedImageError as e:
             logging.error(f"UnidentifiedImageError: Could not identify image file for xref {xref}: {e}")
             return None
-        
+
         except Exception as e:
             logging.error(f"Unexpected error while extracting image for xref {xref}: {e}")
             return None
 
-    def split_files_across_gpus(self) -> List[List[FileDescriptor]]:
+    def merge_split_with_full_page_indexer(self, files: List[FileDescriptor], num_gpus: int) -> Tuple[
+        List[List[FileDescriptor]], List[List[Tuple[int, str]]]]:
+        """
+        Merge and split files while maintaining a full page index.
+
+        Args:
+            files (List[FileDescriptor]): List of file descriptors to process.
+            num_gpus (int): Number of GPUs to distribute files across.
+
+        Returns:
+            Tuple[List[List[FileDescriptor]], List[List[Tuple[int, str]]]]:
+            File descriptors per GPU and their corresponding page indices.
+        """
+
+        def build_pdf(pdf_pages: List[Tuple[str, Tuple[int, int]]]):
+            """Create a merged PDF from specific page ranges."""
+            output_pdf = fitz.open()
+            for file_path, (lower, upper) in pdf_pages:
+                pdf = fitz.open(file_path)
+                try:
+                    output_pdf.insert_pdf(pdf, lower, upper)
+                except Exception as e:
+                    logger.error(f"Failed to merge file {file_path}: {e}")
+                    continue
+            temp_file = tempfile.NamedTemporaryFile(
+                delete=False, suffix=".pdf"
+            )
+
+            output_pdf.save(temp_file)
+
+            return [FileDescriptor.from_filename(file_path=temp_file.name)]
+
+        def get_total_pages():
+            total = 0
+            delete_index = []
+            for i, descriptor in enumerate(files):
+                try:
+                    total += len(fitz.open(descriptor.file_path))
+                except Exception as e:
+                    logger.error(f"Failed to open pdf at {descriptor.file_path}: {e}")
+                    delete_index.append(i)
+            return total, delete_index
+
+        total_pages, delete_index = get_total_pages()
+        files = [file for i, file in enumerate(files) if i not in delete_index]
+        full_page_indexer = [list() for _ in range(num_gpus)]
+        pages_per_gpu = (total_pages // num_gpus) + 1
+
+        gpu_file_descriptors = []
+
+        current_gpu_pages = []
+        current_gpu = 0
+        current_num_pages = 0
+
+        for file in files:
+            try:
+                pdf_file = fitz.open(file.file_path)
+            except Exception as e:
+                logger.error(f"Failed to open pdf at {file.file_path}: {e}")
+                continue
+            current_gpu_pages.append((file.file_path, [0, 0]))
+
+            for i, _ in enumerate(pdf_file):
+                if current_num_pages >= pages_per_gpu:
+                    current_gpu_pages[-1][1][1] = i + 1
+                    gpu_file_descriptors.append(build_pdf(current_gpu_pages))
+                    current_gpu += 1
+                    current_num_pages = 0
+                    current_gpu_pages.clear()
+                    current_gpu_pages.append((file.file_path, [i + 2, i + 2]))
+
+                full_page_indexer[current_gpu].append((i, file.file_path))
+                current_num_pages += 1
+
+            current_gpu_pages[-1][1][1] = len(pdf_file)
+
+        if len(current_gpu_pages) > 0:
+            current_gpu_pages[-1][1][1] = i
+            gpu_file_descriptors.append(build_pdf(current_gpu_pages))
+
+        return gpu_file_descriptors, full_page_indexer
+
+    def _split_files_across_gpus(self) -> List[List[FileDescriptor]]: #TODO CALL THAT SHIT
         """
         Split the files across GPUs.
         """
-        chunked_files, self.full_page_indexer = merge_split_with_full_page_indexer(self.files, torch.cuda.device_count())
+        chunked_files, self.full_page_indexer = self.merge_split_with_full_page_indexer(self.files, torch.cuda.device_count())
         return chunked_files
 
-    @classmethod
-    def get_file_len(cls, file: FileDescriptor) -> int:
+    def get_file_len(self, file: FileDescriptor) -> int:
         try:
             pdf_doc = fitz.open(file.file_path)
             return len(pdf_doc)
@@ -301,3 +378,81 @@ class PDFProcessor(Processor):
                 f"Error while trying to get the number of pages of the PDF file {file.file_path}. Error: {str(e)}"
             )
             return -1
+
+    def _consolidate_modalities(self, result: ProcessorResult) -> ProcessorResult:
+        """
+        Consolidates modalities in the processing result by updating file paths and copying files 
+        (e.g., images) to a specified output directory.
+
+        Args:
+            result (ProcessorResult): The result containing samples with modalities to be consolidated.
+
+        Returns:
+            ProcessorResult: A new ProcessorResult object with updated modalities paths.
+        """
+
+        def save_new_image(image_path: str, output_folder_path: str) -> str | None:
+            """
+            Saves a copy of the image to a new location.
+
+            Args:
+                image_path (str): The original path of the image.
+                output_folder_path (str): The destination folder to save the image.
+
+            Returns:
+                Optional[str]: The new image path or None.
+            """
+            try:
+                image = Image.open(image_path)
+                new_path = os.path.join(output_folder_path, "images", os.path.basename(image_path))
+                image.save(new_path)
+                return new_path
+            except Exception as e:
+                logger.error(f"Failed to save image {image_path if image_path else 'Unknown'}: {str(e)}")
+                return None
+
+        # Create a new list of samples with updated modalities paths
+        new_samples = []
+        output_path = self.config.custom_config.get("output_path", "output")
+        os.makedirs(os.path.join(output_path, "images"), exist_ok=True)
+
+        for sample in result.samples:
+            old_modalities = sample.modalities
+            new_modalities = []
+            for modality in old_modalities:
+                modality_type = modality.type
+                modality_value = modality.value
+
+                if modality_type == "image":
+                    new_modality_value = save_new_image(modality_value, output_path)
+
+                    # Only add the modality if the new path is valid
+                    if new_modality_value:
+                        new_modalities.append(MultimodalRawInput(modality_type, new_modality_value))
+                    else:
+                        logger.warning(f"Skipping invalid or failed image modality: {modality_value}")
+                else:
+                    # Keep non-image modalities unchanged
+                    new_modalities.append(modality)
+
+            new_sample = MultimodalSample(
+                sample.text,
+                new_modalities,
+                sample.metadata
+            )
+            new_samples.append(new_sample)
+        # We have a list of ProcessorResult, we need to merge them
+        return ProcessorResult(new_samples)
+
+
+
+    def create_sample_list_already_saved_images(texts: List[str], images: List[List[str]], paths: List[str]) -> List[dict]:
+        def create_sample_alread_saved_images(texts: List[str], images: List[str], file_path: str) -> dict:
+            return {
+                "text": "\n".join(texts),
+                "modalities": [{"type": "image", "value": img} for img in images],
+                "metadata": {"file_path": file_path},
+            }
+
+        return [create_sample_alread_saved_images([text], image, path) for text, image, path in zip(texts, images, paths)]
+
