+++ src/mmore/process/processors/pptx_processor.py
@@ -1,11 +1,13 @@
 import logging
 import io
+
+from PIL.Image import Image
 from pptx import Presentation
 from pptx.enum.shapes import MSO_SHAPE_TYPE
 from PIL import Image
-from src.mmore.process.utils import clean_text, create_sample, clean_image
+from src.mmore.process.utils import clean_text, clean_image
 from src.mmore.type import FileDescriptor
-from .processor import Processor, ProcessorConfig
+from .processor import Processor, ProcessorConfig, ProcessorResult
 from typing import List, Dict, Any
 
 logger = logging.getLogger(__name__)
@@ -19,6 +21,7 @@ class PPTXProcessor(Processor):
         files (List[FileDescriptor]): List of files to be processed.
         config (ProcessorConfig): Configuration for the processor.
     """
+
     def __init__(self, files, config=None):
         """
         Args:
@@ -27,8 +30,7 @@ class PPTXProcessor(Processor):
         """
         super().__init__(files, config=config or ProcessorConfig())
 
-    @classmethod
-    def accepts(cls, file: FileDescriptor) -> bool:
+    def accepts(self, file: FileDescriptor) -> bool:
         """
         Args:
             file (FileDescriptor): The file descriptor to check.
@@ -43,44 +45,9 @@ class PPTXProcessor(Processor):
         Returns:
             tuple: A tuple (False, False) indicating no GPU requirement for both standard and fast modes.
         """
-        return False, False
+        return False
 
-    def _extract_slide_content(
-            self, slide, all_text: List[str], embedded_images: List[Image.Image]
-    ):
-        """
-        Extract text and images from a slide and append them to the provided lists.
-
-        Args:
-            slide (Slide): A slide from the PowerPoint presentation.
-            all_text (List[str]): List to store extracted text.
-            embedded_images (List[Image.Image]): List to store extracted images.
-        """
-        # Sort shapes by their vertical position
-        shape_list = sorted(
-            (shape for shape in slide.shapes if hasattr(shape, "top")),
-            key=lambda s: s.top,
-        )
-
-        for shape in shape_list:
-            # Extract text from shape
-            if shape.has_text_frame:
-                cleaned_text = clean_text(shape.text)
-                if cleaned_text.strip():
-                    all_text.append(cleaned_text)
-
-            # Extract images from shape
-            if shape.shape_type == MSO_SHAPE_TYPE.PICTURE:
-                try:
-                    pil_image = Image.open(io.BytesIO(shape.image.blob)).convert("RGBA")
-                    if clean_image(pil_image):
-                        embedded_images.append(pil_image)
-                        all_text.append(self.config.attachment_tag)
-
-                except Exception as e:
-                    logger.error(f"Failed to extract image from slide: {e}")
-
-    def process_implementation(self, file_path: str) -> Dict[str, Any]:
+    def process_one_file(self, file_path: str, fast: bool = False) -> ProcessorResult:
         """
         Process a single PPTX file. Extracts text, images, and notes from each slide.
 
@@ -93,22 +60,46 @@ class PPTXProcessor(Processor):
         The method processes each slide, extracting text and images from shapes,
         and extracts notes if present. The elements are sorted by their vertical position.
         """
+        super().process_one_file(file_path, fast=fast)
+
         logger.info(f"Processing PowerPoint file: {file_path}")
         try:
             prs = Presentation(file_path)
         except Exception as e:
             logger.error(f"Failed to open PowerPoint file {file_path}: {e}")
-            return {"text": "", "modalities": []}
+            return self.create_sample([], [], file_path)
 
-        all_text = []
-        embedded_images = []
+        all_text: list[str] = []
+        embedded_images: list[Image.Image] = []
 
         try:
             for slide in prs.slides:
-                # Extract text and images from each shape on the slide
-                self._extract_slide_content(slide, all_text, embedded_images)
+                # 1) Extract text and images from slide
+                # Sort shapes by their vertical position
+                shape_list = sorted(
+                    (shape for shape in slide.shapes if hasattr(shape, "top")),
+                    key=lambda s: s.top,
+                )
 
-                # Extract text from slide notes if present
+                for shape in shape_list:
+                    # Extract text from shape
+                    if shape.has_text_frame:
+                        cleaned_text = clean_text(shape.text)
+                        if cleaned_text.strip():
+                            all_text.append(cleaned_text)
+
+                    # Extract images from shape
+                    if shape.shape_type == MSO_SHAPE_TYPE.PICTURE:
+                        try:
+                            pil_image = Image.open(io.BytesIO(shape.image.blob)).convert("RGBA")
+                            if clean_image(pil_image):
+                                embedded_images.append(pil_image)
+                                all_text.append(self.config.attachment_tag)
+
+                        except Exception as e:
+                            logger.error(f"Failed to extract image from slide: {e}")
+
+                # 2) Extract text from slide notes if present
                 if slide.has_notes_slide and slide.notes_slide.notes_text_frame:
                     notes = slide.notes_slide.notes_text_frame
                     for paragraph in notes.paragraphs:
@@ -120,4 +111,4 @@ class PPTXProcessor(Processor):
         except Exception as e:
             logger.error(f"[PPTX] Error processing slides in {file_path}: {e}")
 
-        return create_sample(all_text, embedded_images, file_path)
+        return self.create_sample(all_text, embedded_images, file_path)
